Implement failing jobs from standard jobs table
